Revert "[REFACTOR] DimmedFullscreen -> DimmedOverlay 로 변경"
This reverts commit fb26bb26f249c075467c03d5a690a1a352bc561d.

# Conflicts:
#	Packy/Screens/OpenBox/BoxDetail/BoxDetailPhotoView.swift
