+++ Packy/Screens/MainTab/MainTabView.swift
@@ -12,7 +12,7 @@ import ComposableArchitecture
 
 struct MainTabView: View {
     private let store: StoreOf<MainTabFeature>
-    @ObservedObject private(set) var viewStore: ViewStoreOf<MainTabFeature>
+    @ObservedObject private var viewStore: ViewStoreOf<MainTabFeature>
     @State var presentBottomSheet: Bool = false
 
     init(store: StoreOf<MainTabFeature>) {
@@ -110,16 +110,17 @@ struct MainTabView: View {
 
 private extension MainTabView {
     var content: some View {
-        archiveOverlay {
-            VStack(spacing: 0) {
-                tabView
-                bottomTabBar
-            }
+        VStack(spacing: 0) {
+            tabView
+            bottomTabBar
         }
         .bottomSheet(
             isPresented: .init(
                 get: { viewStore.popupBox.popupBox != nil },
-                set: { if $0 == false { viewStore.send(.popupBox(._hideBottomSheet)) } }
+                set: {
+                    guard $0 == false else { return }
+                    viewStore.send(.popupBox(._hideBottomSheet))
+                }
             ),
             detents: [.height(525)]
         ) {
